+++ src/Opulence/Ioc/Bootstrappers/Factories/BootstrapperRegistryFactory.php
@@ -13,7 +13,6 @@ namespace Opulence\Ioc\Bootstrappers\Factories;
 use Opulence\Ioc\Bootstrappers\BootstrapperRegistry;
 use Opulence\Ioc\Bootstrappers\IBootstrapperRegistry;
 use Opulence\Ioc\Bootstrappers\IBootstrapperResolver;
-use Opulence\Ioc\Bootstrappers\ILazyBootstrapper;
 
 /**
  * Defines the bootstrapper registry factory
@@ -40,14 +39,7 @@ class BootstrapperRegistryFactory implements IBootstrapperRegistryFactory
         $bootstrapperRegistry = new BootstrapperRegistry();
 
         foreach ($bootstrapperObjects as $bootstrapperObject) {
-            if ($bootstrapperObject instanceof ILazyBootstrapper) {
-                $bootstrapperRegistry->registerLazyBootstrapper(
-                    $bootstrapperObject->getBindings(),
-                    get_class($bootstrapperObject)
-                );
-            } else {
-                $bootstrapperRegistry->registerEagerBootstrapper(get_class($bootstrapperObject));
-            }
+            $bootstrapperRegistry->registerBootstrapper($bootstrapperObject);
         }
 
         return $bootstrapperRegistry;